refact: EditCardActivity에서 setContent 부분 분리
