+++ app/src/main/java/nextstep/payments/ui/editcard/EditCardScreen.kt
@@ -8,16 +8,25 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import kotlinx.coroutines.flow.collectLatest
 import nextstep.payments.R
 import nextstep.payments.model.card.CardNumber
+import nextstep.payments.model.card.CreditCard
 import nextstep.payments.ui.component.card.CardBankInformation
 import nextstep.payments.ui.component.card.CardInformation
 import nextstep.payments.ui.component.card.edit.EditableCardPage
@@ -25,6 +34,55 @@ import nextstep.payments.ui.theme.PaymentsTheme
 import java.time.Month
 import java.time.YearMonth
 import java.util.UUID
+
+@Composable
+internal fun EditCardScreen(
+    viewModel: EditCardViewModel,
+    onCardUpdated: () -> Unit,
+    onBackClick: () -> Unit,
+) {
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    val context = LocalContext.current
+    val targetCardInformation = viewModel.targetCard.toCardInformation()
+    val snackbarHostState = remember { SnackbarHostState() }
+
+    LaunchedEffect(true) {
+        viewModel.errorFlow.collectLatest {
+            snackbarHostState.showSnackbar(message = context.getString(R.string.card_edit_error))
+        }
+    }
+    LaunchedEffect(true) {
+        viewModel.cardUpdated.collectLatest { onCardUpdated() }
+    }
+    Scaffold(
+        snackbarHost = { SnackbarHost(hostState = snackbarHostState) },
+        content = { contentPadding ->
+            EditCardScreen(
+                cardNumber = uiState.cardNumber,
+                expirationDate = uiState.expirationDate,
+                ownerName = uiState.ownerName,
+                password = uiState.password,
+                cardInformation = targetCardInformation,
+                setCardNumber = viewModel::setCardNumber,
+                setExpiredDate = viewModel::setExpiredDate,
+                setOwnerName = viewModel::setOwnerName,
+                setPassword = viewModel::setPassword,
+                onBackClick = onBackClick,
+                onSaveClick = viewModel::updateCard,
+                modifier = Modifier.padding(contentPadding),
+            )
+        }
+    )
+}
+
+private fun CreditCard.toCardInformation(): CardInformation = CardInformation(
+    id = this.id,
+    numberFirst = this.cardNumbers[0],
+    numberSecond = this.cardNumbers[1],
+    ownerName = this.ownerName,
+    expirationDate = this.expirationDate,
+    bank = CardBankInformation.from(this.bankType)
+)
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
